requirements.txt
PyMuPDF>=1.24.0
Pillow>=10.0.0
requests>=2.28.0
python-docx>=0.8.11
openpyxl>=3.1.0
faster-whisper>=1.0.0
imageio-ffmpeg>=0.4.0
PyYAML>=6.0
tqdm>=4.60.0
